style(dialog): fix lint errors
DSY-793
