+++ src/components/Dialog/index.ts
@@ -1,1 +1,4 @@
 export { Dialog } from './Dialog';
+export {
+  DialogActions, DialogContent, DialogContentText, DialogTitle,
+} from './Dialog.styles';
